==> notebooks/02_Training.ipynb <==
from sklearn.ensemble import GradientBoostingClassifier
import pandas as pd
from sklearn.model_selection import KFold
import pickle
from sklearn.metrics import classification_report
from sklearn.compose import ColumnTransformer
from sklearn.impute import SimpleImputer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import RepeatedStratifiedKFold, train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from xgboost.sklearn import XGBClassifier

def train_and_save_xgb():
    """Train the XGBBoost and save the model - Loads the pipeline, Retrieve the processed data       
     .. WARNING::  Be careful to not tie knots when moving the snake!       
     :param int x: The x position where move to.        
     :param int y: The y position where move to.        
    """
    pipeline = pickle.load(open('../models/pipe.pkl', 'rb'))
    train = pd.read_csv("../data/processed/application_train.csv")
    int_train = pd.read_csv("../data/interim/train_before_normalisation.csv")
    X_train = train#.iloc[:, 1:240]
    y_train = int_train.TARGET
    #X_train = train.drop('TARGET',axis=1)#pd.concat([train, test], ignore_index=True).drop("TARGET", axis=1)
    print("training the gradient booster...")
    pipeline.named_steps['clf'].fit(train, y_train)
    print("getting the best params...")
    gs = GridSearchCV(pipeline.named_steps["clf"], {"max_depth": [1, 1.3, 1.5]}, n_jobs=-1, cv=5, scoring="accuracy")
    gs.fit(train, y_train)
    print("best params : ",gs.best_params_['max_depth'])
    pipeline.set_params(clf__max_depth=gs.best_params_['max_depth'])
    print('re training the gradient booster...')
    pipeline.named_steps['clf'].fit(train, y_train)
    pickle.dump(pipeline, open('../models/pipe.pkl','wb'))

train_and_save_xgb()
