==> notebooks/03_Predict.ipynb <==
import pickle
import pandas as pd
from sklearn.metrics import accuracy_score, classification_report

pipe = pickle.load(open('../models/pipe.pkl', 'rb'))
train = pd.read_csv("../data/processed/application_train.csv")#.set_index('SK_ID_CURR')
X_test = pd.read_csv("../data/processed/application_test.csv")#.set_index('SK_ID_CURR')
int_train = pd.read_csv("../data/interim/train_before_normalisation.csv")
y = int_train.TARGET

#y_pred = pipe.named_steps["clf"].predict(X_test)
test = pd.read_csv("../data/raw/application_test.csv")
test['prediction'] = pd.Series(y_pred)
test.to_csv("../data/output/predictions.csv")

print('accuracy_score : ',accuracy_score(y, y_pred))
print(classification_report(y, y_pred)) 

pd.set_option('display.max_columns', None)

Mean Accuracy

clf.score(X_test, y_test)
